+++ CHANGELOG.md
@@ -4,3 +4,11 @@ Changelog for `gpustat`
 ## v0.1.1
 
 - Fix the display format 
+
+## v0.1.3
+
+- Add driver version
+- Add Encoder/Decoder utilization rates
+- Add fan speed
+- Remove unused user option
+- Update dependencies